refactor: App.tsxをpagesディレクトリ下に移動
ページ関連ファイルをまとめるため
