+++ frontend/src/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import App from "./App";
+import App from "./pages/App";
 import reportWebVitals from "./reportWebVitals";
 import SignIn from "./pages/SignIn";
 import TeamsSelect from "./pages/TeamsSelect";
