+++ frontend/src/test/App.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
-import App from "../App";
+import App from "../pages/App";
 
 test("renders learn react link", () => {
   render(
