+++ frontend/src/test/TeamsShow.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import renderer from 'react-test-renderer'; // eslint-disable-line import/no-extraneous-dependencies
+import renderer from "react-test-renderer"; // eslint-disable-line import/no-extraneous-dependencies
 import { BrowserRouter } from "react-router-dom";
 import TeamsShow from "../pages/TeamsShow";
